+++ packages/skeleton/postcss.config.cjs
@@ -1,3 +1,4 @@
+const path = require('path');
 const tailwindcss = require('tailwindcss');
 const autoprefixer = require('autoprefixer');
 
@@ -5,7 +6,7 @@ const config = {
 	plugins: [
 		require('postcss-import'),
 		//Some plugins, like tailwindcss/nesting, need to run before Tailwind,
-		tailwindcss(),
+		tailwindcss(path.resolve(__dirname, './tailwind.config.js')),
 		//But others, like autoprefixer, need to run after,
 		autoprefixer
 	]