+++ sites/skeleton.dev/postcss.config.cjs
@@ -1,6 +1,7 @@
+const path = require('path');
+const tailwindcss = require('tailwindcss');
+const autoprefixer = require('autoprefixer');
+
 module.exports = {
-	plugins: {
-		tailwindcss: {},
-		autoprefixer: {}
-	}
+	plugins: [tailwindcss(path.resolve(__dirname, './tailwind.config.js')), autoprefixer]
 };